refactor(e2e): split steps per feature and fill in COVERAGE.md gaps
common.ts had grown to 41 steps in 307 lines, mixing navigation, learn,
and quiz concerns. Split into three files so each one only owns the
steps that are conceptually shared across that scope:

  common.ts (80 lines)      page Givens, generic link/tab clicks,
                            URL Thens, generic visibility assertions,
                            BeforeScenario coverage hook
  learn.steps.ts (75)       notebook detail Givens, story/word-card/
                            mindmap navigation, origin/node Thens
  quiz.steps.ts (154)       mode/notebook selection, answer inputs,
                            submit/continue/finish, prompt/card Thens

COVERAGE.md was stale: /learn/[id] was marked uncovered (the story
reader scenario added it), the etymology Scenario Outline was rewritten
as three separate scenarios, and several routes/interactions were
missing entirely. Rewrite the route table to list every scenario and
add per-route interaction lists for /learn/[id], /quiz/etymology-*
(separately), and the Include-unstudied / story-row clicks the
existing scenarios already cover. Consolidate untested cases under a
single "Known gaps" backlog.
